LayoutTests/imported/w3c/web-platform-tests/html/semantics/invokers/invokeelement-interface.tentative-expected.txt



PASS invokeTargetElement reflects invokee HTML element
PASS invokeTargetElement reflects set value
PASS invokeTargetElement reflects set value across shadow root into light dom
PASS invokeTargetElement does not reflect set value inside shadowroot
PASS invokeTargetElement throws error on assignment of non Element
FAIL invokeAction reflects '' when attribute not present assert_equals: expected "" but got "auto"
FAIL invokeAction reflects '' when attribute empty, setAttribute version assert_equals: expected "" but got "auto"
PASS invokeAction reflects same casing
FAIL invokeAction reflects '' when attribute empty, IDL version assert_equals: expected "" but got "auto"
PASS invokeAction reflects tostring value
FAIL invokeAction reflects '' when attribute set to [] assert_equals: expected "" but got "auto"
PASS invokeAction reflects tostring value 2
